requirements.txt
ws4py
pycrypto
cherrypy
ldap3
requests
python-gammu
flask
flask-login
flask-sqlalchemy
timeout-decorator
bcrypt
configparser
